Add data structure for 2D points
This doesn't use the PIMPL idiom since I don't expect the interface to change at all for this struct unless the coordinate type is changed in the future.
The PIMPL idiom would incur unnecessary indirection because it has to dereference that pointer for every access of its memory or any function. That would kill performance here.