Add function to check collisions between convex polygons
This is my initial implementation completely from the top of my head and from theory. I've never run it yet so I wonder if it works correctly. It's a fairly simple algorithm but I applied some optimisations which are intuitively true but I couldn't prove from the top of my head.
It should be sufficient to project only the vertices of the other polygon and see if they are on the right hand side of the edge. This polygon is always on the left, but not infinitely so, so in theory there could be a separating axis where the other polygon is completely on the left and so far that they are separated on that axis. However my intuition says that there is then always another edge on the opposite side of this polygon where they are actually separated around the 0-projection. I couldn't prove that solidly though.
If I implement this algorithm in Apex I'll have to go the distance to prove that in the documentation.