Move headers into a convack subfolder
This is good practice for C++ applications. This way, when someone wants to use the headers of this library, the headers can be included with #include <convack/header.hpp> rather than just #include <header.hpp>, so the global include namespace is not muddled. So this way people can safely include the Convack include directory in their include directories.