Fix getting the left-most point of a set of points
If there are three points with the same X coordinate and they are all three the left-most points, we don't want to start with the middle vertex. This middle vertex should not be in the convex hull since it's bisecting two colinear edges then. The result contains a useless vertex then (or even doesn't terminate depending on the implementation of the rest of the gift wrapping algorithm). This makes it take the one with the lowest Y coordinate which is never the middle one. Could as well have been the lowest Y coordinate, that doesn't matter.