Fix comparing how far left or right multiple points are
We need to compare them directly with is_left to see if one vertex is left of a line through the other vertex, rather than seeing how far the both of them are left compared to an arbitrary other line. The output of is_left can't be compared fairly since it's not normalised and depends on the vector sizes involved.