Add test for getting convex hull around two triangles
This is the simplest happy path there is, really. Two tiny polygons, not intersecting or anything. It actually also doesn't do the binary search in this case because it's either vertex 0 or it happens to need the pivot point. So more extensive testing is required, maybe with some more high-resolution models.