Give each convex polygon a unique identifier
Increment it only when constructing a new convex polygon from a fixed list of vertices, not when copying a convex polygon.
This is necessary to identify convex polygons during the beam search algorithm. The beam search gets a fixed (global) list of convex polygons. From this list, each candidate copies convex polygons which it then transforms. The transformed copies need to be identifiable as being the same convex polygons as what was already in that global list.
I considered an actual UUID first (random). This would maybe make it possible to generate them multi-threadedly. However I realised that generating this multi-threaded is really not going to be necessary in the near future. I want to implement a multi-threaded search, but the important UUIDs are only the ones generated before the search starts, in the global list of convex polygons. So those are generated single-threaded by the user of the library. It'll generate UUIDs for intermediary results, such as the convex hulls around convex polygons necessary to calculate the score, but those are not used at all. It's not crucial that the UUIDs are unique across all threads then. So then this simple incrementing is sufficient, even if the increment is not atomic.